create single source for common maths methods for mri and jruby
