+++ lib/nmatrix/cruby/math.rb
@@ -51,7 +51,23 @@ class NMatrix
   #   - +StorageTypeError+ -> ATLAS functions only work on dense matrices.
   #
   def getrf!
-    ipiv = LUDecomposition.new(self.twoDMat).getPivot.to_a
+    raise(StorageTypeError, "ATLAS functions only work on dense matrices") unless self.dense?
+
+    #For row-major matrices, clapack_getrf uses a different convention than
+    #described above (U has unit diagonal elements instead of L and columns
+    #are interchanged rather than rows). For column-major matrices, clapack
+    #uses the stanard conventions. So we just transpose the matrix before
+    #and after calling clapack_getrf.
+    #Unfortunately, this is not a very good way, uses a lot of memory.
+    temp = self.transpose
+    ipiv = NMatrix::LAPACK::clapack_getrf(:col, self.shape[0], self.shape[1], temp, self.shape[0])
+    temp = temp.transpose
+    self[0...self.shape[0], 0...self.shape[1]] = temp
+
+    #for some reason, in clapack_getrf, the indices in ipiv start from 0
+    #instead of 1 as in LAPACK.
+    ipiv.each_index { |i| ipiv[i]+=1 }
+
     return ipiv
   end
 
@@ -178,18 +194,7 @@ class NMatrix
   #
   def potrf!(which)
     # The real implementation is in the plugin files.
-    cholesky = CholeskyDecomposition.new(self.twoDMat)
-    if which == :upper
-      u = create_dummy_nmatrix
-      twoDMat = cholesky.getLT
-      u.s = ArrayRealVector.new(ArrayGenerator.getArrayDouble(twoDMat.getData, @shape[0], @shape[1]))
-      return u
-    else
-      l = create_dummy_nmatrix
-      twoDMat = cholesky.getL
-      l.s = ArrayRealVector.new(ArrayGenerator.getArrayDouble(twoDMat.getData, @shape[0], @shape[1]))
-      return l
-    end
+    raise(NotImplementedError, "potrf! requires either the nmatrix-atlas or nmatrix-lapacke gem")
   end
 
   def potrf_upper!
@@ -213,15 +218,10 @@ class NMatrix
   # symmetric or Hermitian. However, it is still your responsibility to make
   # sure it is positive-definite.
   def factorize_cholesky
-    # raise "Matrix must be symmetric/Hermitian for Cholesky factorization" unless self.hermitian?
-    cholesky = CholeskyDecomposition.new(self.twoDMat)
-    l = create_dummy_nmatrix
-    twoDMat = cholesky.getL
-    l.s = ArrayRealVector.new(ArrayGenerator.getArrayDouble(twoDMat.getData, @shape[0], @shape[1]))
-    u = create_dummy_nmatrix
-    twoDMat = cholesky.getLT
-    u.s = ArrayRealVector.new(ArrayGenerator.getArrayDouble(twoDMat.getData, @shape[0], @shape[1]))
-    return [u,l]
+    raise "Matrix must be symmetric/Hermitian for Cholesky factorization" unless self.hermitian?
+    l = self.clone.potrf_lower!.tril!
+    u = l.conjugate_transpose
+    [u,l]
   end
 
   #
@@ -240,11 +240,12 @@ class NMatrix
   def factorize_lu with_permutation_matrix=nil
     raise(NotImplementedError, "only implemented for dense storage") unless self.stype == :dense
     raise(NotImplementedError, "matrix is not 2-dimensional") unless self.dimensions == 2
-    t = self.clone
-    pivot = create_dummy_nmatrix
-    twoDMat = LUDecomposition.new(self.twoDMat).getP
-    pivot.s = ArrayRealVector.new(ArrayGenerator.getArrayDouble(twoDMat.getData, @shape[0], @shape[1]))
-    return [t,pivot]
+
+    t     = self.clone
+    pivot = t.getrf!
+    return t unless with_permutation_matrix
+
+    [t, FactorizeLUMethods.permutation_matrix_from(pivot)]
   end
 
   #
@@ -265,20 +266,26 @@ class NMatrix
   #   - +ShapeError+ -> Input must be a 2-dimensional matrix to have a QR decomposition.
   #
   def factorize_qr
-
     raise(NotImplementedError, "only implemented for dense storage") unless self.stype == :dense
     raise(ShapeError, "Input must be a 2-dimensional matrix to have a QR decomposition") unless self.dim == 2
-    qrdecomp = QRDecomposition.new(self.twoDMat)
 
-    qmat = create_dummy_nmatrix
-    qtwoDMat = qrdecomp.getQ
-    qmat.s = ArrayRealVector.new(ArrayGenerator.getArrayDouble(qtwoDMat.getData, @shape[0], @shape[1]))
+    rows, columns = self.shape
+    r = self.clone
+    tau =  r.geqrf!
 
-    rmat = create_dummy_nmatrix
-    rtwoDMat = qrdecomp.getR
-    rmat.s = ArrayRealVector.new(ArrayGenerator.getArrayDouble(rtwoDMat.getData, @shape[0], @shape[1]))
-    return [qmat,rmat]
+    #Obtain Q
+    q = self.complex_dtype? ? r.unmqr(tau) : r.ormqr(tau)
 
+    #Obtain R
+    if rows <= columns
+      r.upper_triangle!
+    #Need to account for upper trapezoidal structure if R is a tall rectangle (rows > columns)
+    else
+      r[0...columns, 0...columns].upper_triangle!
+      r[columns...rows, 0...columns] = 0
+    end
+
+    [q,r]
   end
 
   # Solve the matrix equation AX = B, where A is +self+, B is the first
@@ -337,21 +344,243 @@ class NMatrix
     n    = self.shape[0]
     nrhs = b.shape[1]
 
-    nmatrix = create_dummy_nmatrix
     case opts[:form]
-    when :general, :upper_tri, :upper_triangular, :lower_tri, :lower_triangular
-      #LU solver
-      solver = LUDecomposition.new(self.twoDMat).getSolver
-      nmatrix.s = solver.solve(b.s)
-      return nmatrix
+    when :general
+      clone = self.clone
+      ipiv = NMatrix::LAPACK.clapack_getrf(:row, n, n, clone, n)
+      # When we call clapack_getrs with :row, actually only the first matrix
+      # (i.e. clone) is interpreted as row-major, while the other matrix (x)
+      # is interpreted as column-major. See here: http://math-atlas.sourceforge.net/faq.html#RowSolve
+      # So we must transpose x before and after
+      # calling it.
+      x = x.transpose
+      NMatrix::LAPACK.clapack_getrs(:row, :no_transpose, n, nrhs, clone, n, ipiv, x, n)
+      x.transpose
+    when :upper_tri, :upper_triangular
+      raise(ArgumentError, "upper triangular solver does not work with complex dtypes") if
+        complex_dtype? or b.complex_dtype?
+      # this is the correct function call; see https://github.com/SciRuby/nmatrix/issues/374
+      NMatrix::BLAS::cblas_trsm(:row, :left, :upper, false, :nounit, n, nrhs, 1.0, self, n, x, nrhs)
+      x
+    when :lower_tri, :lower_triangular
+      raise(ArgumentError, "lower triangular solver does not work with complex dtypes") if
+        complex_dtype? or b.complex_dtype?
+      NMatrix::BLAS::cblas_trsm(:row, :left, :lower, false, :nounit, n, nrhs, 1.0, self, n, x, nrhs)
+      x
     when :pos_def, :positive_definite
-      solver = CholeskyDecomposition.new(self.twoDMat).getSolver
-      nmatrix.s = solver.solve(b.s)
-      return nmatrix
+      u, l = self.factorize_cholesky
+      z = l.solve(b, form: :lower_tri)
+      u.solve(z, form: :upper_tri)
     else
       raise(ArgumentError, "#{opts[:form]} is not a valid form option")
     end
+  end
 
+  #
+  # call-seq:
+  #     least_squares(b) -> NMatrix
+  #     least_squares(b, tolerance: 10e-10) -> NMatrix
+  #
+  # Provides the linear least squares approximation of an under-determined system
+  # using QR factorization provided that the matrix is not rank-deficient.
+  #
+  # Only works for dense matrices.
+  #
+  # * *Arguments* :
+  #   - +b+ -> The solution column vector NMatrix of A * X = b.
+  #   - +tolerance:+ -> Absolute tolerance to check if a diagonal element in A = QR is near 0
+  #
+  # * *Returns* :
+  #   - NMatrix that is a column vector with the LLS solution
+  #
+  # * *Raises* :
+  #   - +ArgumentError+ -> least squares approximation only works for non-complex types
+  #   - +ShapeError+ -> system must be under-determined ( rows > columns )
+  #
+  # Examples :-
+  #
+  #   a = NMatrix.new([3,2], [2.0, 0, -1, 1, 0, 2])
+  #
+  #   b = NMatrix.new([3,1], [1.0, 0, -1])
+  #
+  #   a.least_squares(b)
+  #     =>[
+  #         [ 0.33333333333333326 ]
+  #         [ -0.3333333333333334 ]
+  #       ]
+  #
+  def least_squares(b, tolerance: 10e-6)
+    raise(ArgumentError, "least squares approximation only works for non-complex types") if
+      self.complex_dtype?
+
+    rows, columns = self.shape
+
+    raise(ShapeError, "system must be under-determined ( rows > columns )") unless
+      rows > columns
+
+    #Perform economical QR factorization
+    r = self.clone
+    tau = r.geqrf!
+    q_transpose_b = r.ormqr(tau, :left, :transpose, b)
+
+    #Obtain R from geqrf! intermediate
+    r[0...columns, 0...columns].upper_triangle!
+    r[columns...rows, 0...columns] = 0
+
+    diagonal = r.diagonal
+
+    raise(ArgumentError, "rank deficient matrix") if diagonal.any? { |x| x == 0 }
+
+    if diagonal.any? { |x| x.abs < tolerance }
+      warn "warning: A diagonal element of R in A = QR is close to zero ;" <<
+           " indicates a possible loss of precision"
+    end
+
+    # Transform the system A * X = B to R1 * X = B2 where B2 = Q1_t * B
+    r1 = r[0...columns, 0...columns]
+    b2 = q_transpose_b[0...columns]
+
+    nrhs = b2.shape[1]
+
+    #Solve the upper triangular system
+    NMatrix::BLAS::cblas_trsm(:row, :left, :upper, false, :nounit, r1.shape[0], nrhs, 1.0, r1, r1.shape[0], b2, nrhs)
+    b2
+  end
+
+  #
+  # call-seq:
+  #     gesvd! -> [u, sigma, v_transpose]
+  #     gesvd! -> [u, sigma, v_conjugate_transpose] # complex
+  #
+  # Compute the singular value decomposition of a matrix using LAPACK's GESVD function.
+  # This is destructive, modifying the source NMatrix.  See also #gesdd.
+  #
+  # Optionally accepts a +workspace_size+ parameter, which will be honored only if it is larger than what LAPACK
+  # requires.
+  #
+  def gesvd!(workspace_size=1)
+    NMatrix::LAPACK::gesvd(self, workspace_size)
+  end
+
+  #
+  # call-seq:
+  #     gesvd -> [u, sigma, v_transpose]
+  #     gesvd -> [u, sigma, v_conjugate_transpose] # complex
+  #
+  # Compute the singular value decomposition of a matrix using LAPACK's GESVD function.
+  #
+  # Optionally accepts a +workspace_size+ parameter, which will be honored only if it is larger than what LAPACK
+  # requires.
+  #
+  def gesvd(workspace_size=1)
+    self.clone.gesvd!(workspace_size)
+  end
+
+
+
+  #
+  # call-seq:
+  #     gesdd! -> [u, sigma, v_transpose]
+  #     gesdd! -> [u, sigma, v_conjugate_transpose] # complex
+  #
+  # Compute the singular value decomposition of a matrix using LAPACK's GESDD function. This uses a divide-and-conquer
+  # strategy. This is destructive, modifying the source NMatrix.  See also #gesvd.
+  #
+  # Optionally accepts a +workspace_size+ parameter, which will be honored only if it is larger than what LAPACK
+  # requires.
+  #
+  def gesdd!(workspace_size=nil)
+    NMatrix::LAPACK::gesdd(self, workspace_size)
+  end
+
+  #
+  # call-seq:
+  #     gesdd -> [u, sigma, v_transpose]
+  #     gesdd -> [u, sigma, v_conjugate_transpose] # complex
+  #
+  # Compute the singular value decomposition of a matrix using LAPACK's GESDD function. This uses a divide-and-conquer
+  # strategy. See also #gesvd.
+  #
+  # Optionally accepts a +workspace_size+ parameter, which will be honored only if it is larger than what LAPACK
+  # requires.
+  #
+  def gesdd(workspace_size=nil)
+    self.clone.gesdd!(workspace_size)
+  end
+
+  #
+  # call-seq:
+  #     laswp!(ary) -> NMatrix
+  #
+  # In-place permute the columns of a dense matrix using LASWP according to the order given as an array +ary+.
+  #
+  # If +:convention+ is +:lapack+, then +ary+ represents a sequence of pair-wise permutations which are
+  # performed successively. That is, the i'th entry of +ary+ is the index of the column to swap
+  # the i'th column with, having already applied all earlier swaps.
+  #
+  # If +:convention+ is +:intuitive+, then +ary+ represents the order of columns after the permutation.
+  # That is, the i'th entry of +ary+ is the index of the column that will be in position i after the
+  # reordering (Matlab-like behaviour). This is the default.
+  #
+  # Not yet implemented for yale or list.
+  #
+  # == Arguments
+  #
+  # * +ary+ - An Array specifying the order of the columns. See above for details.
+  #
+  # == Options
+  #
+  # * +:covention+ - Possible values are +:lapack+ and +:intuitive+. Default is +:intuitive+. See above for details.
+  #
+  def laswp!(ary, opts={})
+    raise(StorageTypeError, "ATLAS functions only work on dense matrices") unless self.dense?
+    opts = { convention: :intuitive }.merge(opts)
+
+    if opts[:convention] == :intuitive
+      if ary.length != ary.uniq.length
+        raise(ArgumentError, "No duplicated entries in the order array are allowed under convention :intuitive")
+      end
+      n = self.shape[1]
+      p = []
+      order = (0...n).to_a
+      0.upto(n-2) do |i|
+        p[i] = order.index(ary[i])
+        order[i], order[p[i]] = order[p[i]], order[i]
+      end
+      p[n-1] = n-1
+    else
+      p = ary
+    end
+
+    NMatrix::LAPACK::laswp(self, p)
+  end
+
+  #
+  # call-seq:
+  #     laswp(ary) -> NMatrix
+  #
+  # Permute the columns of a dense matrix using LASWP according to the order given in an array +ary+.
+  #
+  # If +:convention+ is +:lapack+, then +ary+ represents a sequence of pair-wise permutations which are
+  # performed successively. That is, the i'th entry of +ary+ is the index of the column to swap
+  # the i'th column with, having already applied all earlier swaps. This is the default.
+  #
+  # If +:convention+ is +:intuitive+, then +ary+ represents the order of columns after the permutation.
+  # That is, the i'th entry of +ary+ is the index of the column that will be in position i after the
+  # reordering (Matlab-like behaviour).
+  #
+  # Not yet implemented for yale or list.
+  #
+  # == Arguments
+  #
+  # * +ary+ - An Array specifying the order of the columns. See above for details.
+  #
+  # == Options
+  #
+  # * +:covention+ - Possible values are +:lapack+ and +:intuitive+. Default is +:lapack+. See above for details.
+  #
+  def laswp(ary, opts={})
+    self.clone.laswp!(ary, opts)
   end
 
   #
@@ -381,7 +610,27 @@ class NMatrix
   #
   def det
     raise(ShapeError, "determinant can be calculated only for square matrices") unless self.dim == 2 && self.shape[0] == self.shape[1]
-    self.det_exact2
+
+    # Cast to a dtype for which getrf is implemented
+    new_dtype = self.integer_dtype? ? :float64 : self.dtype
+    copy = self.cast(:dense, new_dtype)
+
+    # Need to know the number of permutations. We'll add up the diagonals of
+    # the factorized matrix.
+    pivot = copy.getrf!
+
+    num_perm = 0 #number of permutations
+    pivot.each_with_index do |swap, i|
+      #pivot indexes rows starting from 1, instead of 0, so need to subtract 1 here
+      num_perm += 1 if swap-1 != i
+    end
+    prod = num_perm % 2 == 1 ? -1 : 1 # odd permutations => negative
+    [shape[0],shape[1]].min.times do |i|
+      prod *= copy[i,i]
+    end
+
+    # Convert back to an integer if necessary
+    new_dtype != self.dtype ? prod.round : prod #prevent rounding errors
   end
 
   #
@@ -448,7 +697,8 @@ class NMatrix
   #
   # Return the 2-norm of the vector. This is the BLAS nrm2 routine.
   def nrm2 incx=1, n=nil
-    self.twoDMat.getFrobeniusNorm()
+    return method_missing(:nrm2, incx, n) unless vector?
+    NMatrix::BLAS::nrm2(self, incx, self.size / incx)
   end
   alias :norm2 :nrm2
 
@@ -465,13 +715,12 @@ class NMatrix
   # Return the scaling result of the matrix. BLAS scal will be invoked if provided.
 
   def scale!(alpha, incx=1, n=nil)
-    #FIXME
-    # raise(DataTypeError, "Incompatible data type for the scaling factor") unless
-    #     NMatrix::upcast(self.dtype, NMatrix::min_dtype(alpha)) == self.dtype
-    raise(DataTypeError, "Incompatible data type for the scaling factor") if
-        self.dtype == :int8
-    @s.mapMultiplyToSelf(alpha)
-    return self
+    raise(DataTypeError, "Incompatible data type for the scaling factor") unless
+        NMatrix::upcast(self.dtype, NMatrix::min_dtype(alpha)) == self.dtype
+    return NMatrix::BLAS::scal(alpha, self, incx, self.size / incx) if NMatrix::BLAS.method_defined? :scal
+    self.each_stored_with_indices do |e, *i|
+      self[*i] = e*alpha
+    end
   end
 
   #
@@ -486,16 +735,10 @@ class NMatrix
   # Return the scaling result of the matrix. BLAS scal will be invoked if provided.
 
   def scale(alpha, incx=1, n=nil)
-    # FIXME
-    # raise(DataTypeError, "Incompatible data type for the scaling factor") unless
-    #     NMatrix::upcast(self.dtype, NMatrix::min_dtype(alpha)) == self.dtype
-    raise(DataTypeError, "Incompatible data type for the scaling factor") if
-        self.dtype == :byte || self.dtype == :int8 || self.dtype == :int16 ||
-        self.dtype == :int32 || self.dtype == :int64
-    nmatrix = NMatrix.new :copy
-    nmatrix.shape = @shape.clone
-    nmatrix.s = ArrayRealVector.new(@s.toArray.clone).mapMultiplyToSelf(alpha)
-    return nmatrix
+    return self.clone.scale!(alpha, incx, n)
   end
+
+  alias :permute_columns  :laswp
+  alias :permute_columns! :laswp!
 
 end
